# api/openapi.yaml
openapi: 3.0.3
info:
  title: riskline API
  description: |
    Open-source AI risk classification — CLI and API, deterministic and auditable,
    not an LLM guessing. Advisory tool only; not legal advice.

    P0 ships eu-ai-act (hard law) and nist-ai-rmf (mapping). Request regimes[] to
    select packs; geographic_scope is not a regime selector and must not be overloaded.
  version: 0.5.0-alpha
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: New World Coder
    url: https://github.com/new-world-coder/riskline
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /v1/classify:
    post:
      operationId: classifySystem
      summary: Classify an AI system under selected regime packs
      description: |
        Accepts a structured description of an AI system and returns a risk tier,
        clause references, plain-language rationale, and recommended controls.
        When regimes is omitted, the server default applies (typically eu-ai-act).
        Top-level fields always reflect the first resolved regime for backward
        compatibility; classifications[] appears when multiple regimes are evaluated.
      tags:
        - Classification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassifyRequest'
            examples:
              recruiting_screener:
                summary: CV screening assistant
                value:
                  name: "Hiring Assist"
                  purpose: "Screen job applicants and rank candidates for interview"
                  data_types:
                    - personal_data
                    - employment_data
                  deployment_context: saas_b2b
                  autonomy_level: decision_support
                  affected_population: job_applicants
                  geographic_scope: eu
      responses:
        '200':
          description: Classification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassifyResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/diff:
    post:
      operationId: diffSystemDescription
      summary: Detect material change between two system descriptions
      description: |
        Compares baseline and current ClassifyRequest values and returns whether
        the change is material for conformity (re-classify vs re-assure only).
      tags:
        - Assurance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiffRequest'
      responses:
        '200':
          description: Material change analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialChangeResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/assure:
    post:
      operationId: assureControls
      summary: Verify technical controls and compute conformity state
      description: |
        Layer 2 Assure: given a classification and probe results keyed by
        technical_hook (or control id), returns GREEN/AMBER/RED conformity state
        and SHA-256 evidence records (local hash chain, not blockchain).
      tags:
        - Assurance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssureRequest'
      responses:
        '200':
          description: Assurance result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssureResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/runtime/register:
    post:
      operationId: registerRuntime
      summary: Register a deployment baseline and derived runtime policy
      description: |
        Layer 3 Runtime: binds a classified and assured system to a deployment,
        producing a baseline fingerprint and policy envelope for continuous verification.
        Handler not implemented in Phase 1 — schema and contract only.
      tags:
        - Runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRuntimeRequest'
      responses:
        '200':
          description: Registration result with baseline and policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterRuntimeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Not implemented (Phase 1 schema stub)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runtime/observe:
    post:
      operationId: observeRuntime
      summary: Submit a metadata-only runtime observation
      description: |
        Accepts a runtime signal (inference, tool call, deploy, etc.) for
        continuous verification. Handler not implemented in Phase 1.
      tags:
        - Runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeObservation'
      responses:
        '202':
          description: Observation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObserveRuntimeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Not implemented (Phase 1 schema stub)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runtime/verify:
    post:
      operationId: verifyRuntime
      summary: Verify a runtime observation against policy and baseline
      description: |
        Compares an observation to registered or inline policy/baseline and returns
        conformity state, violations, and a verification receipt stub.
        Handler not implemented in Phase 1.
      tags:
        - Runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRuntimeRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyRuntimeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Not implemented (Phase 1 schema stub)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ClassifyRequest:
      type: object
      required:
        - purpose
        - data_types
        - deployment_context
        - autonomy_level
        - affected_population
      properties:
        name:
          type: string
          description: Optional human label for the system
          maxLength: 200
        purpose:
          type: string
          description: What the AI system is used for
          minLength: 1
          maxLength: 4000
        data_types:
          type: array
          description: Categories of data the system processes
          minItems: 1
          items:
            $ref: '#/components/schemas/DataType'
        deployment_context:
          $ref: '#/components/schemas/DeploymentContext'
        autonomy_level:
          $ref: '#/components/schemas/AutonomyLevel'
        affected_population:
          $ref: '#/components/schemas/AffectedPopulation'
        geographic_scope:
          $ref: '#/components/schemas/GeographicScope'
          default: eu
          description: |
            Hint about where the system is placed on the market / who it affects.
            Not a regime pack selector — use regimes[] for that.
        regimes:
          type: array
          description: |
            Regime pack IDs to evaluate (e.g. eu-ai-act, nist-ai-rmf). Empty/omitted
            uses the server or CLI default. Unknown IDs are rejected.
          items:
            type: string
            example: eu-ai-act
          uniqueItems: true
        uses_biometric_categorisation:
          type: boolean
          default: false
        real_time_remote_biometric_id:
          type: boolean
          default: false
        social_scoring:
          type: boolean
          default: false
        emotion_recognition_workplace_or_education:
          type: boolean
          default: false
        manipulative_techniques:
          type: boolean
          default: false
        exploits_vulnerabilities:
          type: boolean
          default: false
        model_id:
          type: string
          description: Deployed LLM/model version identifier (runtime metadata for material-change detection)
        system_prompt_hash:
          type: string
          description: Hash of the current system prompt (runtime metadata)
        tools:
          type: array
          description: Connected tool/API names (runtime metadata)
          items:
            type: string
        human_approval_required:
          type: boolean
          description: Whether human-in-the-loop is active (runtime metadata; change alone triggers reassure)
          default: false
    DataType:
      type: string
      enum:
        - personal_data
        - special_category_data
        - biometric_data
        - employment_data
        - financial_credit_data
        - education_data
        - health_data
        - law_enforcement_data
        - publicly_available
        - synthetic_or_non_personal
        - other
    DeploymentContext:
      type: string
      enum:
        - saas_b2b
        - saas_b2c
        - on_prem
        - embedded_product
        - public_authority
        - law_enforcement
        - internal_only
        - other
    AutonomyLevel:
      type: string
      enum:
        - content_generation
        - decision_support
        - automated_decision
        - autonomous_action
    AffectedPopulation:
      type: string
      enum:
        - general_public
        - customers
        - employees
        - job_applicants
        - students
        - patients
        - credit_consumers
        - suspects_or_accused
        - children
        - other
    GeographicScope:
      type: string
      enum:
        - eu
        - eu_and_global
        - us
        - us_and_global
        - non_eu
        - unknown
    RiskTier:
      type: string
      enum:
        - prohibited
        - high_risk
        - limited_risk
        - minimal_risk
    ClassifyResponse:
      type: object
      required:
        - risk_tier
        - ruleset_version
        - last_updated
        - matched_rules
        - rationale
        - recommended_controls
        - disclaimer
      properties:
        name:
          type: string
        risk_tier:
          $ref: '#/components/schemas/RiskTier'
        ruleset_version:
          type: string
          example: "eu-ai-act-2024-v0.1.0"
        last_updated:
          type: string
          format: date
          example: "2026-08-04"
        matched_rules:
          type: array
          items:
            $ref: '#/components/schemas/MatchedRule'
        rationale:
          type: string
        recommended_controls:
          type: array
          items:
            type: string
        technical_controls:
          type: array
          items:
            $ref: '#/components/schemas/TechnicalControl'
        judgment_calls:
          type: array
          description: Boundary cases where the ruleset made an explicit judgment call
          items:
            type: string
        disclaimer:
          type: string
          example: >-
            This classification is an advisory tool based on a versioned ruleset.
            It is not legal advice and is not a substitute for qualified counsel
            or a formal conformity assessment.
        regime:
          type: string
          description: |
            Pack that produced the top-level fields. Omitted on the historical
            single-pack eu-ai-act default response for backward compatibility.
          example: eu-ai-act
        mapping_only:
          type: boolean
          description: |
            True when the primary regime is a mapping pack (e.g. nist-ai-rmf).
            risk_tier is not a US legal conclusion — inspect technical_controls.
        classifications:
          type: array
          description: |
            Per-regime outcomes. Present only when more than one regime was
            evaluated. Other jurisdictions must not be collapsed into EU tiers
            at the top level — inspect character and regime-specific fields here
            as packs beyond eu-ai-act ship.
          items:
            $ref: '#/components/schemas/RegimeClassification'
    RegimeClassification:
      type: object
      required:
        - regime
        - character
        - risk_tier
        - ruleset_version
        - last_updated
        - matched_rules
        - rationale
        - recommended_controls
      properties:
        regime:
          type: string
          example: eu-ai-act
        character:
          type: string
          description: Legal character of this pack's outcomes
          enum:
            - hard_law
            - principles
            - guidance
            - mapping
        risk_tier:
          $ref: '#/components/schemas/RiskTier'
        ruleset_version:
          type: string
        last_updated:
          type: string
          format: date
        matched_rules:
          type: array
          items:
            $ref: '#/components/schemas/MatchedRule'
        rationale:
          type: string
        recommended_controls:
          type: array
          items:
            type: string
        technical_controls:
          type: array
          items:
            $ref: '#/components/schemas/TechnicalControl'
        judgment_calls:
          type: array
          items:
            type: string
        mapping_only:
          type: boolean
          description: True for mapping packs — risk_tier is not a legal tier.
    TechnicalControl:
      type: object
      required:
        - id
        - paper_ref
        - summary
        - technical_hook
        - evidence_type
      properties:
        id:
          type: string
          example: human-review-employment
        paper_ref:
          type: string
          example: "NIST AI RMF Manage 2.3 / CPPA ADMT (significant decision)"
        summary:
          type: string
        technical_hook:
          type: string
          example: workflow.human_review_required
        evidence_type:
          type: string
          example: audit_log + signed_classify_json
    MatchedRule:
      type: object
      required:
        - id
        - tier
        - article_or_annex
        - summary
        - ruleset_version
        - last_updated
      properties:
        id:
          type: string
        tier:
          $ref: '#/components/schemas/RiskTier'
        article_or_annex:
          type: string
          example: "Annex III (4)(a)"
        summary:
          type: string
        ruleset_version:
          type: string
        last_updated:
          type: string
          format: date
    DiffRequest:
      type: object
      required:
        - baseline
        - current
      properties:
        baseline:
          $ref: '#/components/schemas/ClassifyRequest'
        current:
          $ref: '#/components/schemas/ClassifyRequest'
    ConformityState:
      type: string
      enum:
        - green
        - amber
        - red
    ConformityImpact:
      type: string
      enum:
        - none
        - reassure
        - reclassify
    MaterialChangeResult:
      type: object
      required:
        - material
        - conformity_impact
        - prior_fingerprint
        - current_fingerprint
        - summary
        - disclaimer
      properties:
        material:
          type: boolean
        conformity_impact:
          $ref: '#/components/schemas/ConformityImpact'
        changed_fields:
          type: array
          items:
            type: string
        prior_fingerprint:
          type: string
        current_fingerprint:
          type: string
        summary:
          type: string
        disclaimer:
          type: string
    ControlVerdict:
      type: object
      required:
        - control_id
        - technical_hook
        - paper_ref
        - evidence_type
        - passed
      properties:
        control_id:
          type: string
        technical_hook:
          type: string
        paper_ref:
          type: string
        evidence_type:
          type: string
        passed:
          type: boolean
        reason:
          type: string
    EvidenceRecord:
      type: object
      required:
        - id
        - timestamp
        - ruleset_version
        - control_id
        - result
        - content_hash
        - algorithm
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        system_name:
          type: string
        ruleset_version:
          type: string
        control_id:
          type: string
        result:
          type: string
        content_hash:
          type: string
        previous_hash:
          type: string
        algorithm:
          type: string
    AssureRequest:
      type: object
      required:
        - classification
        - probes
      properties:
        classification:
          $ref: '#/components/schemas/ClassifyResponse'
        probes:
          type: object
          additionalProperties:
            type: boolean
        previous_hash:
          type: string
    AssureResponse:
      type: object
      required:
        - conformity_state
        - control_verdicts
        - evidence_records
        - unverified_count
        - failed_count
        - summary
        - disclaimer
      properties:
        conformity_state:
          $ref: '#/components/schemas/ConformityState'
        control_verdicts:
          type: array
          items:
            $ref: '#/components/schemas/ControlVerdict'
        evidence_records:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRecord'
        unverified_count:
          type: integer
        failed_count:
          type: integer
        summary:
          type: string
        disclaimer:
          type: string
...
    RuntimeEventType:
      type: string
      enum:
        - inference
        - tool_call
        - human_override
        - deploy
        - config_change
    RuntimeObservation:
      type: object
      required:
        - system_id
        - deployment_id
        - event_type
        - human_approval_granted
        - timestamp
      properties:
        system_id:
          type: string
          description: Registered system identifier
        deployment_id:
          type: string
          description: Deployment or environment identifier
        model_id:
          type: string
          description: Deployed model version observed at runtime
        system_prompt_hash:
          type: string
          description: Hash of the active system prompt
        tools:
          type: array
          description: Connected tools observed at runtime
          items:
            type: string
        event_type:
          $ref: '#/components/schemas/RuntimeEventType'
        data_categories:
          type: array
          items:
            $ref: '#/components/schemas/DataType'
        geography:
          $ref: '#/components/schemas/GeographicScope'
        autonomy_level:
          $ref: '#/components/schemas/AutonomyLevel'
        human_approval_granted:
          type: boolean
          description: Whether human approval was active for this event
        policy_version:
          type: string
        ruleset_version:
          type: string
        timestamp:
          type: string
          format: date-time
        metrics:
          type: object
          description: Optional numeric runtime metrics (latency, token counts, etc.)
          additionalProperties:
            type: number
            format: double
    RuntimeBaseline:
      type: object
      required:
        - system_id
        - deployment_id
        - fingerprint
        - approved_at
        - source_classification
      properties:
        system_id:
          type: string
        deployment_id:
          type: string
        fingerprint:
          type: string
          description: SHA-256 fingerprint of the approved runtime posture
        approved_at:
          type: string
          format: date-time
        source_classification:
          $ref: '#/components/schemas/ClassifyRequest'
          description: Classification input fields used to compute the baseline fingerprint
    RuntimePolicy:
      type: object
      required:
        - max_autonomy
        - required_controls
        - risk_tier
        - human_approval_required
      properties:
        approved_models:
          type: array
          items:
            type: string
        forbidden_models:
          type: array
          items:
            type: string
        approved_tools:
          type: array
          items:
            type: string
        forbidden_tools:
          type: array
          items:
            type: string
        approved_data_categories:
          type: array
          items:
            $ref: '#/components/schemas/DataType'
        forbidden_data_categories:
          type: array
          items:
            $ref: '#/components/schemas/DataType'
        approved_geography:
          type: array
          items:
            $ref: '#/components/schemas/GeographicScope'
        forbidden_geography:
          type: array
          items:
            $ref: '#/components/schemas/GeographicScope'
        max_autonomy:
          $ref: '#/components/schemas/AutonomyLevel'
        required_controls:
          type: array
          items:
            type: string
        risk_tier:
          $ref: '#/components/schemas/RiskTier'
        human_approval_required:
          type: boolean
    RuntimeViolationSeverity:
      type: string
      enum:
        - info
        - warn
        - critical
    RuntimeViolation:
      type: object
      required:
        - code
        - severity
        - field
        - expected
        - actual
        - conformity_impact
      properties:
        code:
          type: string
          example: forbidden_model
        severity:
          $ref: '#/components/schemas/RuntimeViolationSeverity'
        field:
          type: string
          example: model_id
        expected:
          type: string
        actual:
          type: string
        conformity_impact:
          $ref: '#/components/schemas/ConformityImpact'
    VerificationResult:
      type: object
      required:
        - conformity_state
        - summary
      properties:
        conformity_state:
          $ref: '#/components/schemas/ConformityState'
        violations:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeViolation'
        risk_delta:
          type: string
          description: Description of risk posture change relative to baseline
        controls_failed:
          type: array
          items:
            type: string
        recommended_action:
          type: string
        summary:
          type: string
    VerificationReceipt:
      type: object
      description: |
        Tamper-evident runtime verification artifact. Local SHA-256 hash chain and
        optional Ed25519 signature — not a public blockchain ledger.
      required:
        - verification_id
        - system_id
        - verified_at
        - runtime_fingerprint
        - baseline_fingerprint
        - policy_version
        - conformity_state
        - receipt_hash
      properties:
        verification_id:
          type: string
        system_id:
          type: string
        verified_at:
          type: string
          format: date-time
        observed_at:
          type: string
          format: date-time
        runtime_fingerprint:
          type: string
        baseline_fingerprint:
          type: string
        policy_version:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeViolation'
        conformity_state:
          $ref: '#/components/schemas/ConformityState'
        receipt_hash:
          type: string
        previous_receipt_hash:
          type: string
        signature:
          type: string
          description: Hex-encoded Ed25519 signature
        public_key:
          type: string
          description: Hex-encoded Ed25519 public key
        algorithm:
          type: string
          enum:
            - ed25519
    RegisterRuntimeRequest:
      type: object
      required:
        - system_id
        - deployment_id
        - source_description
        - classification
        - assure
      properties:
        system_id:
          type: string
        deployment_id:
          type: string
        source_description:
          $ref: '#/components/schemas/ClassifyRequest'
          description: Approved system description used to derive baseline fingerprint and policy
        classification:
          $ref: '#/components/schemas/ClassifyResponse'
        assure:
          $ref: '#/components/schemas/AssureResponse'
        probes:
          type: object
          additionalProperties:
            type: boolean
    RegisterRuntimeResponse:
      type: object
      required:
        - baseline
        - policy
        - disclaimer
      properties:
        baseline:
          $ref: '#/components/schemas/RuntimeBaseline'
        policy:
          $ref: '#/components/schemas/RuntimePolicy'
        disclaimer:
          type: string
    VerifyRuntimeRequest:
      type: object
      required:
        - observation
      properties:
        observation:
          $ref: '#/components/schemas/RuntimeObservation'
        policy:
          $ref: '#/components/schemas/RuntimePolicy'
          description: Optional inline policy; omit to use registered policy
        baseline:
          $ref: '#/components/schemas/RuntimeBaseline'
          description: Optional inline baseline; omit to use registered baseline
    VerifyRuntimeResponse:
      type: object
      required:
        - result
        - receipt
      properties:
        result:
          $ref: '#/components/schemas/VerificationResult'
        receipt:
          $ref: '#/components/schemas/VerificationReceipt'
    ObserveRuntimeResponse:
      type: object
      required:
        - accepted
      properties:
        accepted:
          type: boolean
        message:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - disclaimer
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
        disclaimer:
          type: string